Rename FormActivityConfirmationHelper to SheetActivityConfirmationHelper.
